pyproject.toml
[project]
name = "athena"
version = "0.1.0"
description = "Adversarial Tactical Hearing & Equilibrium Navigation Agent"
requires-python = ">=3.11"
dependencies = [
    "langgraph",
    "langchain-core",
    "langchain-community",
    "mlx-lm",
    "langfuse",
    "pydantic>=2.0",
    "pyyaml",
    "httpx>=0.27",
    "json-repair>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "jsonschema",
]
kg = [
    "graphiti-core>=0.5",
    "neo4j>=5.0",
    "sentence-transformers>=3.0.0",
]
validation = [
    "datasets>=2.0,<3.0",
]

[project.scripts]
athena = "athena.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
